scripts/requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
dbt-core
dbt-postgres
